reset: DB형식에 맞추어 컴포넌트 코드 전체 재작성
