+++ src/components/forms/Button.jsx
@@ -1,10 +1,4 @@
-// Button.jsx
-// 코드 작성자 : GiHhub @whitedev77773
-
-// 외부 라이브러리에서 import
 import styled from 'styled-components';
-
-// 사용자가 생성한 컴포넌트 및 JS파일 import
 
 /**
  * Flat / Outline / Translucent
