+++ src/components/typograph/Text.jsx
@@ -1,11 +1,5 @@
-// Text.jsx
-// 코드 작성자 : GiHhub @whitedev77773
-
-// 외부 라이브러리에서 import
 import React from 'react';
 import styled from 'styled-components';
-
-// 사용자가 생성한 컴포넌트 및 JS파일 import
 
 /**
  * @typedef {"xs" | "s" | "m" | "l" | "xl" | "sxl" | string} Size
@@ -73,6 +67,7 @@ export const Span = styled.span`
 /**
  * Text Component
  * 스타일링된 span을 생성합니다.
+ * 기본 : m, regular
  * @param {IText & IChildren} props
  */
 export const Text = ({
