+++ src/pages/dashboard/Admin.styled.js
@@ -1,4 +1,6 @@
 import styled from 'styled-components';
+
+// 사용자 정의 컴포넌트
 import { Container } from '../../components/forms/Container';
 
 export const AdminListContainer = styled(Container)`
