+++ src/pages/dashboard/History.styled.js
@@ -1,5 +1,7 @@
 import styled from 'styled-components';
-import { Container } from './Dashboard.styled';
+
+// 사용자 정의 컴포넌트
+import { Container } from '../../components/forms/Container';
 
 export const ManageHistory = styled(Container)`
   width: 100%;
