+++ src/pages/dashboard/Home.jsx
@@ -1,21 +1,14 @@
-// Home.jsx
-// 코드 작성자 : GiHhub @whitedev77773
-
-// 외부 라이브러리에서 import
 import { useEffect, useState } from 'react';
 import { useNavigate } from 'react-router-dom';
-import useHome from '../../stores/dashboard/useHome';
 
-// 사용자가 생성한 컴포넌트 및 JS파일 import
-import HomeLoading from './Home.Loading';
-import {
-  Header,
-  Container,
-  ContainerHeader,
-  Button,
-  BoardColumn,
-} from './Dashboard.styled';
+// 사용자 정의 컴포넌트
+import { Header } from './Dashboard.styled';
+import { Container } from '../../components/forms/Container';
+import { ContainerHeader, Button, BoardColumn } from './Home.styled';
 import { NumberDisplay } from '../../components/display/NumberDisplay';
+import HomeLoading from './Home.Loading';
+
+import useHome from '../../stores/dashboard/useHome';
 
 const HistoryBoard = ({ name, data, navigate }) => {
   return (
