+++ src/pages/dashboard/Users.jsx
@@ -1,9 +1,4 @@
-// Users.jsx
-// 코드 작성자 : GiHhub @whitedev77773
-
-// 외부 라이브러리에서 import
-
-// 사용자가 생성한 컴포넌트 및 JS파일 import
+// 사용자 정의 컴포넌트
 import { Header } from './Dashboard.styled';
 
 export default function Users() {
